backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
dataclasses-json==0.6.3
distro==1.9.0
Django==5.0.1
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.8
langsmith==0.0.77
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==0.28.0
packaging==23.2
protobuf==4.25.2
psycopg==3.1.16
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zope.interface==6.1
